Created base dockerfile (#1)
This is used to ensure a consistant build env. It also serves to prevent
pollution of ones own dev env.

Added basic hello world project

Primarily worked on the make file

This make file will be used for all contained projects unless a project
specific makefile is there.

Got makefile working

Dockerfile now makes a symbolic link from makefile.base to makefile
This is so if a project has a makefile it does not clobber the base one
and can use the base make file if need be.

helper script and first docs

Updated README to note scripts dir
Added bash as a supported language to natural docs
refactored project.txt and added the scripts dir to it
updated hello world project to have slightly better docs
Worked on the build_tools.sh script and documented all of its features.